engine-4-onboarding/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12
pyyaml==6.0.2
itsdangerous==2.2.0
